Refactor two provider add buttons to one
