+++ packages/forklift-console-plugin/src/modules/Providers/views/list/components/ProvidersAddButton.tsx
@@ -1,21 +1,31 @@
 import React from 'react';
+import { useHistory } from 'react-router';
+import { getResourceUrl } from 'src/modules/Providers/utils';
+import { useForkliftTranslation } from 'src/utils/i18n';
 
+import { ProviderModelRef } from '@kubev2v/types';
 import { Button } from '@patternfly/react-core';
 
-interface AProvidersAddButtonProps {
-  onClick: () => void;
-  buttonText: string;
-  dataTestId?: string;
-}
-
-export const ProvidersAddButton: React.FC<AProvidersAddButtonProps> = ({
-  onClick,
-  buttonText,
+export const ProvidersAddButton: React.FC<{ namespace: string; dataTestId?: string }> = ({
+  namespace,
   dataTestId,
 }) => {
+  const { t } = useForkliftTranslation();
+  const history = useHistory();
+
+  const providersListURL = getResourceUrl({
+    reference: ProviderModelRef,
+    namespace: namespace,
+    namespaced: namespace !== undefined,
+  });
+
   return (
-    <Button data-testid={dataTestId} variant="primary" onClick={onClick}>
-      {buttonText}
+    <Button
+      data-testid={dataTestId}
+      variant="primary"
+      onClick={() => history.push(`${providersListURL}/~new`)}
+    >
+      {t('Create Provider')}
     </Button>
   );
 };
